Add a new optional authentication protocol.
If the new authentication protocol is used, the tool using the signer will not only get a speaks-for credential, but an authentication token which allows it to ensure that the new credential was just authorized by the user. This only works in the postMessage flow and not in the new PHP/post flow.